back-end/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
dotenv==0.9.9
fastapi==0.115.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.46.2
textual==0.52.1
typing-inspection==0.4.1
typing_extensions==4.14.1

# Git integration
GitPython==3.1.43

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0

# Blockchain
web3==6.15.1
eth-account==0.11.0

# AI/LLM
openai==1.12.0

# YAML support
PyYAML==6.0.1
